// src/app/api/car-logbook/model/logbookFuelEntry.ts
/**
 * Car Logbook Service REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LogbookFuelEntryAverageConsumption } from './logbookFuelEntryAverageConsumption';
import { LogbookFuelEntryPrice } from './logbookFuelEntryPrice';
import { Gasoline } from './gasoline';


export interface LogbookFuelEntry { 
    /**
     * ID
     */
    id?: string;
    date?: string;
    mileage?: number;
    trip?: number;
    units?: number;
    unit?: LogbookFuelEntry.UnitEnum;
    gasoline?: Gasoline;
    comment?: string;
    averageConsumption?: LogbookFuelEntryAverageConsumption;
    price?: LogbookFuelEntryPrice;
    daysSinceLastRefuel?: number;
}
export namespace LogbookFuelEntry {
    export const UnitEnum = {
        Liter: 'LITER',
        Kwh: 'KWH'
    } as const;
    export type UnitEnum = typeof UnitEnum[keyof typeof UnitEnum];
}