// src/app/api/car-logbook/model/logbookFuelEntryPrice.ts
/**
 * Car Logbook Service REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LogbookFuelEntryPrice { 
    per100km?: number;
    perDay?: number;
    perUnit?: number;
    total?: number;
}